reorg and comment to prepare for refactor of read/write
